Attempt to stomp on more id->int conversions in print statements introduced during incorporation of amber improper torsion angles.

+++ src/DuMMForceFieldSubsystem.cpp
@@ -1804,7 +1804,7 @@ void DuMMForceFieldSubsystem::defineBondTorsion
         // Throw an exception if terms for this bond torsion were already defined.
     SimTK_APIARGCHECK4_ALWAYS(ret.second, mm.ApiClassName, MethodName, 
         "bond torsion term(s) were already defined for atom class quad (%d,%d,%d,%d)", 
-        key[0], key[1], key[2], key[3]);
+        (int)key[0], (int)key[1], (int)key[2], (int)key[3]);
 }
 
 // Convenient signature for a bond torsion with only one term.
@@ -1875,7 +1875,7 @@ void DuMMForceFieldSubsystem::defineAmberImproperTorsion
         // Throw an exception if terms for this improper torsion were already defined.
     SimTK_APIARGCHECK4_ALWAYS(ret.second, mm.ApiClassName, MethodName,
         "amber improper torsion term(s) were already defined for atom class quad (%d,%d,%d,%d)",
-        key[0], key[1], key[2], key[3]);
+        (int)key[0], (int)key[1], (int)key[2], (int)key[3]);
 }
 
 // Convenient signature for an amber improper torsion with only one term.
@@ -2573,16 +2573,16 @@ DuMMForceFieldSubsystemRep::getAmberImproperTorsion
    (DuMM::AtomClassId class1, DuMM::AtomClassId class2, DuMM::AtomClassId class3, DuMM::AtomClassId class4) const
 {
 //xxx -> Randy's warning flag
-    printf("aImp--classes: %d-%d-%d-%d\n", class1,
-                                  class2,
-                                  class3,
-                                  class4);
+    printf("aImp--classes: %d-%d-%d-%d\n", (int)class1,
+                                  (int)class2,
+                                  (int)class3,
+                                  (int)class4);
     std::map<AtomClassIdQuad,BondTorsion>::const_iterator i;
     for (i=amberImproperTorsion.begin(); i!=amberImproperTorsion.end(); i++) {
-        printf("aImp-matches: %d-%d-%d-%d\n", i->first[0],
-                                      i->first[1],
-                                      i->first[2],
-                                      i->first[3]);
+        printf( "aImp-matches: %d-%d-%d-%d\n", (int)i->first[0],
+                                      (int)(i->first[1]),
+                                      (int)(i->first[2]),
+                                      (int)(i->first[3]) );
     }
     
     static const AtomClassIdQuad dummyKey(DuMM::InvalidAtomClassId, DuMM::InvalidAtomClassId, DuMM::InvalidAtomClassId, DuMM::InvalidAtomClassId);
